Fix c++1z implementation for type function
This is now based on the rewording of temp.class.spec.match.3
in n4659 which allows for the use of sizeof...(Pack) in non-type
arguments of template specializations if Pack can be deduced.

This makes implementation of TypeFunction and TemplateFunction a lot
easier and cleaner, also improves their debugging.

Note: g++ does not yet support this feature, hence it's only enabled in
clang. In clang, c++1z support is detected via __cpp_template_auto which
is not optimal but fits since it was introduced in the same version of
clang.